bot_telegram/requirements.txt
anthropic>=0.40.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
